Rename all tool identifiers to kebab-case and fix create-document tool
## Tool Naming Updates
- Renamed all tool identifiers from snake_case to kebab-case (e.g., `create_document` → `create-document`)
- Renamed all tool files to match kebab-case convention
- Updated imports in server.ts and test files
- Improved several tool titles for better readability
- Fixed typo in list-documents tool description

## Create Document Tool Fix
- Fixed 501/422 errors by using proper Zod schema with `.passthrough()`
- Added correct API endpoint format with `format=json` parameter
- Implemented proper response handling for 204 No Content status
- Added default document structure with paragraph content
- Updated parameter from `name` to `name` for document ID

## Technical Changes
- All 14 tools now use kebab-case naming convention
- Tests updated to expect new tool names
- Build and test suite passing
- Maintains backward compatibility for API functionality
