+++ src/server.ts
@@ -1,18 +1,18 @@
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
-import registerBatchImportDocuments from './tools/batch_import_documents.js';
-import registerCreateDocument from './tools/create_document.js';
-import registerDeleteDocument from './tools/delete_document.js';
-import registerDuplicateDocument from './tools/duplicate_document.js';
-import registerEncryptDocument from './tools/encrypt_document.js';
-import registerExportMarkdown from './tools/export_markdown.js';
-import registerGetCollaborationHealth from './tools/get_collaboration_health.js';
-import registerGetDocument from './tools/get_document.js';
-import registerGetDocumentStatistics from './tools/get_document_statistics.js';
-import registerGetServerStatistics from './tools/get_server_statistics.js';
-import registerImportMarkdown from './tools/import_markdown.js';
-import registerListDocuments from './tools/list_documents.js';
-import registerSearchDocuments from './tools/search_documents.js';
-import registerUpdateDocument from './tools/update_document.js';
+import registerBatchImportDocuments from './tools/batch-import-documents.js';
+import registerCreateDocument from './tools/create-document.js';
+import registerDeleteDocument from './tools/delete-document.js';
+import registerDuplicateDocument from './tools/duplicate-document.js';
+import registerEncryptDocument from './tools/encrypt-document.js';
+import registerExportMarkdown from './tools/export-markdown.js';
+import registerGetCollaborationHealth from './tools/get-collaboration-health.js';
+import registerGetDocument from './tools/get-document.js';
+import registerGetDocumentStatistics from './tools/get-document-statistics.js';
+import registerGetServerStatistics from './tools/get-server-statistics.js';
+import registerImportMarkdown from './tools/import-markdown.js';
+import registerListDocuments from './tools/list-documents.js';
+import registerSearchDocuments from './tools/search-documents.js';
+import registerUpdateDocument from './tools/update-document.js';
 
 let API_TOKEN: string | undefined;
 let BASE_URL: string | undefined;
