+++ src/tools/batch-import-documents.ts
@@ -1,18 +1,23 @@
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { z } from 'zod';
 
-export default function registerDeleteDocument(
+export default function registerBatchImportDocuments(
   server: McpServer,
   getBaseUrl: () => string,
   getToken: () => string | undefined
 ) {
   server.tool(
-    'delete_document',
-    'Delete a collaborative document',
+    'batch-import-documents',
+    'Bulk import multiple documents using predefined JSON structure',
     {
-      id: z.string().describe('ID of the document to delete'),
+      documents: z.array(z.array(z.object({
+        created_at: z.string().describe('Creation timestamp in ISO format'),
+        version: z.number().describe('Document version number'),
+        name: z.string().describe('Document name/identifier'),
+        tiptap_json: z.object({}).describe('Document content in Tiptap JSON format'),
+      }))).describe('Array of document arrays, where each inner array represents versions of a single document'),
     },
-    async ({ id }) => {
+    async ({ documents }) => {
       try {
         const headers: Record<string, string> = {
           'User-Agent': 'tiptap-collaboration-mcp',
@@ -22,18 +27,19 @@ export default function registerDeleteDocument(
         const token = getToken();
         if (token) headers['Authorization'] = token;
 
-        const response = await fetch(`${getBaseUrl()}/api/documents/${id}`, {
-          method: 'DELETE',
+        const response = await fetch(`${getBaseUrl()}/api/admin/batch-import`, {
+          method: 'PUT',
           headers,
+          body: JSON.stringify(documents),
         });
 
         if (!response.ok) {
-          if (response.status === 404) {
+          if (response.status === 400) {
             return {
               content: [
                 {
                   type: 'text',
-                  text: `Document with ID ${id} not found.`,
+                  text: 'Invalid data provided for batch import. Please check the document structure and format.',
                 },
               ],
             };
@@ -42,7 +48,7 @@ export default function registerDeleteDocument(
             content: [
               {
                 type: 'text',
-                text: `Failed to delete document. HTTP error: ${response.status} ${response.statusText}`,
+                text: `Failed to import documents. HTTP error: ${response.status} ${response.statusText}`,
               },
             ],
           };
@@ -52,7 +58,7 @@ export default function registerDeleteDocument(
           content: [
             {
               type: 'text',
-              text: `Document with ID ${id} deleted successfully.`,
+              text: `Successfully imported ${documents.length} document groups with their versions.`,
             },
           ],
         };
@@ -61,7 +67,7 @@ export default function registerDeleteDocument(
           content: [
             {
               type: 'text',
-              text: `Error deleting document: ${
+              text: `Error importing documents: ${
                 error instanceof Error ? error.message : 'Unknown error'
               }`,
             },
