+++ src/tools/get-document.ts
@@ -1,53 +1,48 @@
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { z } from 'zod';
 
-export default function registerSearchDocuments(
+export default function registerGetDocument(
   server: McpServer,
   getBaseUrl: () => string,
   getToken: () => string | undefined
 ) {
   server.tool(
-    'search_documents',
-    'Search documents using semantic search (requires Tiptap Semantic Search to be enabled)',
+    'get-document',
+    'Get information about a collaborative document',
     {
-      query: z.string().describe('Search query for semantic document search'),
-      limit: z.number().optional().describe('Maximum number of results to return (default: 10)'),
+      id: z.string().describe('ID of the document to retrieve'),
     },
-    async ({ query, limit = 10 }) => {
+    async ({ id }) => {
       try {
         const headers: Record<string, string> = {
           'User-Agent': 'tiptap-collaboration-mcp',
           'Content-Type': 'application/json',
         };
-
+        
         const token = getToken();
         if (token) headers['Authorization'] = token;
 
-        const response = await fetch(`${getBaseUrl()}/api/search`, {
-          method: 'POST',
-          headers,
-          body: JSON.stringify({ query, limit }),
-        });
-
+        const response = await fetch(`${getBaseUrl()}/api/documents/${id}`, { headers });
+        
         if (!response.ok) {
           return {
             content: [
               {
                 type: 'text',
-                text: `Search failed. HTTP error: ${response.status} ${response.statusText}. Note: Semantic search requires Tiptap Semantic Search to be enabled.`,
+                text: `Document with ID ${id} not found. HTTP error: ${response.status} ${response.statusText}`,
               },
             ],
           };
         }
 
-        const searchResults = await response.json();
+        const documentData = await response.json();
 
         return {
           content: [
             {
               type: 'text',
-              text: `Search results for "${query}": ${JSON.stringify(
-                searchResults,
+              text: `Document Information: ${JSON.stringify(
+                documentData,
                 null,
                 2
               )}`,
@@ -59,7 +54,7 @@ export default function registerSearchDocuments(
           content: [
             {
               type: 'text',
-              text: `Error searching documents: ${
+              text: `Error retrieving document: ${
                 error instanceof Error ? error.message : 'Unknown error'
               }`,
             },
