+++ src/tools/update-document.ts
@@ -1,18 +1,20 @@
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { z } from 'zod';
 
-export default function registerGetDocumentStatistics(
+export default function registerUpdateDocument(
   server: McpServer,
   getBaseUrl: () => string,
   getToken: () => string | undefined
 ) {
   server.tool(
-    'get_document_statistics',
-    'Get real-time statistics for a specific document including current connections and connected IPs',
+    'update-document',
+    'Update a collaborative document with new content',
     {
-      id: z.string().describe('ID of the document to get statistics for'),
+      id: z.string().describe('ID of the document to update'),
+      content: z.object({}).describe('Document content in Tiptap JSON format'),
+      mode: z.enum(['replace', 'append']).optional().describe('Update mode: replace entire document or append content (default: replace)'),
     },
-    async ({ id }) => {
+    async ({ id, content, mode = 'replace' }) => {
       try {
         const headers: Record<string, string> = {
           'User-Agent': 'tiptap-collaboration-mcp',
@@ -22,7 +24,11 @@ export default function registerGetDocumentStatistics(
         const token = getToken();
         if (token) headers['Authorization'] = token;
 
-        const response = await fetch(`${getBaseUrl()}/api/documents/${id}/statistics`, { headers });
+        const response = await fetch(`${getBaseUrl()}/api/documents/${id}?mode=${mode}`, {
+          method: 'PATCH',
+          headers,
+          body: JSON.stringify(content),
+        });
 
         if (!response.ok) {
           if (response.status === 404) {
@@ -35,23 +41,31 @@ export default function registerGetDocumentStatistics(
               ],
             };
           }
+          if (response.status === 422) {
+            return {
+              content: [
+                {
+                  type: 'text',
+                  text: `Invalid payload or update cannot be applied to document ${id}.`,
+                },
+              ],
+            };
+          }
           return {
             content: [
               {
                 type: 'text',
-                text: `Failed to retrieve document statistics. HTTP error: ${response.status} ${response.statusText}`,
+                text: `Failed to update document. HTTP error: ${response.status} ${response.statusText}`,
               },
             ],
           };
         }
 
-        const statistics = await response.json();
-
         return {
           content: [
             {
               type: 'text',
-              text: `Document Statistics for ${id}: ${JSON.stringify(statistics, null, 2)}`,
+              text: `Document with ID ${id} updated successfully using ${mode} mode.`,
             },
           ],
         };
@@ -60,7 +74,7 @@ export default function registerGetDocumentStatistics(
           content: [
             {
               type: 'text',
-              text: `Error retrieving document statistics: ${
+              text: `Error updating document: ${
                 error instanceof Error ? error.message : 'Unknown error'
               }`,
             },
